--- docs/source/bmi_calculator.rst
@@ -1,45 +1,0 @@
-bmi\_calculator package
-=======================
-
-Submodules
-----------
-
-bmi\_calculator.calculate\_bmi module
--------------------------------------
-
-.. automodule:: bmi_calculator.calculate_bmi
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-bmi\_calculator.exercise\_plan module
--------------------------------------
-
-.. automodule:: bmi_calculator.exercise_plan
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-bmi\_calculator.project\_bmi module
------------------------------------
-
-.. automodule:: bmi_calculator.project_bmi
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-bmi\_calculator.project\_calories module
-----------------------------------------
-
-.. automodule:: bmi_calculator.project_calories
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: bmi_calculator
-   :members:
-   :undoc-members:
-   :show-inheritance:
