+++ docs/source/bmigraph.rst
@@ -1,0 +1,45 @@
+bmigraph package
+================
+
+Submodules
+----------
+
+bmigraph.calculate\_bmi module
+------------------------------
+
+.. automodule:: bmigraph.calculate_bmi
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+bmigraph.exercise\_plan module
+------------------------------
+
+.. automodule:: bmigraph.exercise_plan
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+bmigraph.project\_bmi module
+----------------------------
+
+.. automodule:: bmigraph.project_bmi
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+bmigraph.project\_calories module
+---------------------------------
+
+.. automodule:: bmigraph.project_calories
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: bmigraph
+   :members:
+   :undoc-members:
+   :show-inheritance:
