+++ tests/bmigraph/test_calculate_bmi.py
@@ -1,4 +1,4 @@
-from bmi_calculator import calculate_bmi
+from bmigraph import calculate_bmi
 import plotly.graph_objects as go
 import pytest
